Bug 407777: Do not include the sym link in teh root files
On MacOS the SymLink launcher should not be included in the root files.
Currently all the other launcher files are excluded, but the sym link
gets packed up (and packed as an actual file). If the Executable
Descriptor includes a sym link (currently only on MacOS), this is added
to the list of excludes.